Move Horaro importing to server
Closes #14
